[issue-8]  Removed dependency on pycrpto and moved to pycryptodome
Problem?

The pycrypto project is no longer maintained and has all sorts
of isses on windows.

Solution!

Switchover to pycryptodome.

-----

Problem?

Blowfish has been deprecrated

Solution!

Upgrade to AES!

-----

Problem?

user-facing tasks often need more granular detail to
be provided to the user as the task progresses, something
that standard tasks seem to be somewhat incapable of
providing.

Solution!

Add detailed backends, tasks, and results that provide
primitves that can be aggregated to provide the user
step-by-step details with respect to what the task
is doing.