[OPS]  Fixed handling of thrown exceptions and direct calling of tasks
Problem?

The DetailedTask failed when called as a direct function

Solution!

Don't save details back to the backend / redis if we are being
used as a direct function and not as a task queued to celery.
When we are called directly, there is no request.id to serve
as a key to the redis backend.