[OPS]  Fixed up config bunch get errors
Problem?

Using get on a config bunch did not work consistently.

Solution!

Fix the way we used the pieces in the get function.
Even though we had walked to the parent, we stupidly insisted
on asking for the whole key instead of only for the last
piece of the key.